fix: make sure creation identity page state are synchronised
The PagerView has its own internal state for the different pages and can be controlled by a ref.selectPage() method but the currently selected page is not available so a componnent state is needed. It appears some bug can desynchronise the two "current page" states. There was already a useEffect to update the PagerView selected page from the component state "currentPage". This commit also add a callback to update the component "currentPage" state if the PagerView somehow changed.